Test: PointStream 의 동작 여부와 실행 순서에 대한 테스트를 추가했습니다.
- 비동기 함수의 종료를 대기하기 위해 Awaitility 를 사용했으나 최선의 방법인지는 조금 망설여짐
- charge, use 할 때 유효성 검사를 위해 IdentityMap 을 사용했는데, 이는 결국 스케일 아웃을 통한 해결이기에 더 나은 개선방법(알고리즘으로 개선)에 대해 고민해야 할 듯
